fix(ci): 让 L 系列完成态通过全平台质量门
远端 GCC、Clang tidy 与 ARM64/QEMU 暴露了本地原生构建看不到的确定性、静态分析和仿真性能门问题。本提交保持既有语义，显式扩宽指针偏移运算、合并重复分支、从生成器消除绑定代码重复，并让调试源位置继续受同一 seqlock 版本窗口保护。ARM64 交叉测试仍执行质量扫描并报告耗时，只不再用 QEMU 开销判定原生 100ms 性能目标。

Constraint: L∀ 要求 Windows/Linux/ARM64、clang-tidy、生成同步与调试快照一致性同时成立

Rejected: 降低 clang-tidy 全局规则 | 会掩盖本轮真实诊断

Rejected: ARM64 排除整个质量测试 | 会丢失功能与零分配验证

Confidence: high

Scope-risk: moderate

Directive: 生成绑定只能从 generate_st_binding_pin_catalog.py 修改；调试发布字段必须留在 seqlock 版本窗口内读取

Tested: Windows Release 90/90；WSL clang-tidy 81 TU 全扫后根因定点复验；ARM64/QEMU L0 quality；RT scan；replay 18/2409；bindings 134/1476；feature-set 12/174/0；conformance 62

Not-tested: 本提交推送后的远端完整 CI

+++ core/st/configuration_runtime.h
@@ -159,7 +159,8 @@ public:
                     return load_failed();
             }
             runtime.write_stamps.assign(
-                (runtime.image_program.process_image.output_bytes +
+                (static_cast<std::size_t>(
+                     runtime.image_program.process_image.output_bytes) +
                  runtime.image_program.process_image.memory_bytes) * 8U,
                 0);
             runtime.write_owners.assign(runtime.write_stamps.size(), 0);
@@ -660,7 +661,8 @@ private:
                 runtime.image_program.initial_data.insert(
                     runtime.image_program.initial_data.end(),
                     program.initial_data.begin() + var.var_offset,
-                    program.initial_data.begin() + required);
+                    program.initial_data.begin() +
+                        static_cast<std::ptrdiff_t>(required));
             }
         }
         return true;
